Consolidated functions that are part of a vRA 7 SDK into one single file (api_client.go).
We are working towards separating the SDK out from the rest of the plugin functionality.
