+++ vrealize/api_client.go
@@ -1,6 +1,7 @@
 package vrealize
 
 import (
+	"encoding/json"
 	"crypto/tls"
 	"fmt"
 	"log"
@@ -95,4 +96,132 @@ func (c *APIClient) Authenticate() error {
 
 	//Return true on success
 	return nil
+}
+
+//DestroyMachine - To set resource destroy call
+func (vRAClient *APIClient) DestroyMachine(destroyTemplate *ActionTemplate, resourceViewTemplate *ResourceView) (*ActionResponseTemplate, error) {
+	//Get a destroy template URL from given resource template
+	var destroyactionURL string
+	destroyactionURL = getactionURL(resourceViewTemplate, "POST: {com.vmware.csp.component.cafe.composition@resource.action.deployment.destroy.name}")
+	//Raise an error if any exception raised while fetching delete resource URL
+	if len(destroyactionURL) == 0 {
+		return nil, fmt.Errorf("Resource is not created or not found")
+	}
+
+	actionResponse := new(ActionResponseTemplate)
+	apiError := new(APIError)
+
+	//Set a REST call with delete resource request and delete resource template as a data
+	resp, err := vRAClient.HTTPClient.New().Post(destroyactionURL).
+		BodyJSON(destroyTemplate).Receive(actionResponse, apiError)
+
+	if resp.StatusCode != 201 {
+		return nil, err
+	}
+
+	if !apiError.isEmpty() {
+		return nil, apiError
+	}
+
+	return actionResponse, nil
+}
+
+//PowerOffMachine - To set resource power-off call
+func (vRAClient *APIClient) PowerOffMachine(powerOffTemplate *ActionTemplate, resourceViewTemplate *ResourceView) (*ActionResponseTemplate, error) {
+	//Get power-off resource URL from given template
+	var powerOffMachineactionURL string
+	powerOffMachineactionURL = getactionURL(resourceViewTemplate, "POST: {com.vmware.csp.component.iaas.proxy.provider@resource.action.name.machine.PowerOff}")
+	//Raise an exception if error got while fetching URL
+	if len(powerOffMachineactionURL) == 0 {
+		return nil, fmt.Errorf("resource is not created or not found")
+	}
+
+	actionResponse := new(ActionResponseTemplate)
+	apiError := new(APIError)
+
+	//Set a rest call to power-off the resource with resource power-off template as a data
+	response, err := vRAClient.HTTPClient.New().Post(powerOffMachineactionURL).
+		BodyJSON(powerOffTemplate).Receive(actionResponse, apiError)
+
+	response.Close = true
+	if response.StatusCode == 201 {
+		return actionResponse, nil
+	}
+
+	if !apiError.isEmpty() {
+		return nil, apiError
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	return nil, err
+}
+
+//GetRequestStatus - To read request status of resource
+// which is used to show information to user post create call.
+func (vRAClient *APIClient) GetRequestStatus(ResourceID string) (*RequestStatusView, error) {
+	//Form a URL to read request status
+	path := fmt.Sprintf("catalog-service/api/consumer/requests/%s", ResourceID)
+	RequestStatusViewTemplate := new(RequestStatusView)
+	apiError := new(APIError)
+	//Set a REST call and fetch a resource request status
+	_, err := vRAClient.HTTPClient.New().Get(path).Receive(RequestStatusViewTemplate, apiError)
+	if err != nil {
+		return nil, err
+	}
+	if !apiError.isEmpty() {
+		return nil, apiError
+	}
+	return RequestStatusViewTemplate, nil
+}
+
+// GetDeploymentState - Read the state of a vRA7 Deployment
+func (vRAClient *APIClient) GetDeploymentState(CatalogRequestId string) (*ResourceView, error) {
+	//Form an URL to fetch resource list view
+	path := fmt.Sprintf("catalog-service/api/consumer/requests/%s"+
+		"/resourceViews", CatalogRequestId)
+	ResourceView := new(ResourceView)
+	apiError := new(APIError)
+	//Set a REST call to fetch resource view data
+	_, err := vRAClient.HTTPClient.New().Get(path).Receive(ResourceView, apiError)
+	if err != nil {
+		return nil, err
+	}
+	if !apiError.isEmpty() {
+		return nil, apiError
+	}
+	return ResourceView, nil
+}
+
+//RequestMachine - To set create resource REST call
+func (vRAClient *APIClient) RequestMachine(template *CatalogItemTemplate) (*RequestMachineResponse, error) {
+	//Form a path to set a REST call to create a machine
+	path := fmt.Sprintf("/catalog-service/api/consumer/entitledCatalogItems/%s"+
+		"/requests", template.CatalogItemID)
+
+	requestMachineRes := new(RequestMachineResponse)
+	apiError := new(APIError)
+
+	jsonBody, jErr := json.Marshal(template)
+	if jErr != nil {
+		log.Printf("Error marshalling template as JSON")
+		return nil, jErr
+	}
+
+	log.Printf("JSON Request Info: %s", jsonBody)
+	//Set a REST call to create a machine
+	_, err := vRAClient.HTTPClient.New().Post(path).BodyJSON(template).
+		Receive(requestMachineRes, apiError)
+
+	if err != nil {
+		return nil, err
+	}
+
+	if !apiError.isEmpty() {
+		return nil, apiError
+	}
+
+	return requestMachineRes, nil
 }
